requirements.txt
argcomplete==3.2.1
asgiref==3.7.2
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
cffi==1.16.0
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
defusedxml==0.7.1
Django==5.0
django-allauth==0.59.0
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
durationpy==0.6
hjson==3.1.0
idna==3.6
inflection==0.5.1
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.3
oauthlib==3.2.2
packaging==23.2
placebo==0.9.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tqdm==4.66.1
troposphere==4.5.3
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==3.0.1
zappa==0.58.0
